model_training/requirements.txt
tensorflow
keras
numpy
pandas
opencv-python
requests
Pillow
scikit-learn
